+++ main.py
@@ -1,3 +1,3 @@
 import scripts.taptap as taptap
 
-apk_path = taptap.TapTapClient().get_latest()
+apk_path = taptap.TapTapClient().get_apk()